vmm: refactor vmcs caching mechanism
Currently, the "if" path in vmcs_flush() would not be executed since the
dirty_count is initialized as VMCS_FIELDS_COUNT and no chance to clear it.

To fix the issue and simplify the code logic, the dirty_count/dirty_fields
and related caching logic will be removed in this patch.

Change-Id: Ib17ce8799604ebd4ca5edf7f8083961de1efce80
Signed-off-by: Qi Yadong <yadong.qi@intel.com>